requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiohttp-swagger==1.0.16
aiosignal==1.3.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
cryptography==41.0.3
dill==0.3.7
distro==1.8.0
dnacentersdk==2.6.6
fastjsonschema==2.18.0
frozenlist==1.4.0
future==0.18.3
genie==22.11
genie.libs.clean==22.11
genie.libs.conf==22.11
genie.libs.filetransferutils==22.11
genie.libs.health==22.11
genie.libs.ops==22.11
genie.libs.parser==22.11
genie.libs.sdk==22.11
gitdb==4.0.10
GitPython==3.1.18
grpcio==1.57.0
idna==3.4
Jinja2==3.1.2
jsonpickle==3.0.2
junit-xml==1.9
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
ncclient==0.6.13
netaddr==0.8.0
numpy==1.24.4
packaging==23.1
pandas==2.0.3
paramiko==3.3.1
pathspec==0.11.2
prettytable==3.8.0
protobuf==4.24.0
psutil==5.9.5
pyats==22.11
pyats.aereport==22.11
pyats.aetest==22.11
pyats.async==22.11
pyats.connections==22.11
pyats.contrib==22.11
pyats.datastructures==22.11
pyats.easypy==22.11
pyats.kleenex==22.11
pyats.log==22.11
pyats.reporter==22.11
pyats.results==22.11
pyats.tcl==22.11
pyats.topology==22.11
pyats.utils==22.11
pycparser==2.21
pyftpdlib==1.5.7
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
tftpy==0.8.0
tqdm==4.66.1
typing-extensions==4.7.1
tzdata==2023.3
unicon==22.11
unicon.plugins==22.11
urllib3==2.0.4
wcwidth==0.2.6
xlrd==1.2.0
XlsxWriter==3.1.2
xlwt==1.3.0
xmltodict==0.13.0
yamllint==1.32.0
yang.connector==23.7
yarl==1.9.2